+++ HISTORY.md
@@ -1,8 +1,10 @@
 
+## 2022-03-28
+  * Add VariableBuffer process
 
 ## 2022-01-19
   * Add RasterSurfaceArea process
-  * 
+  
 ## 2021-02-16
   * Update LinearLineOfSight, RadialLineOfSight processes